+++ changelog.md
@@ -6,6 +6,7 @@
 
 - Added a marker in the WordPress backend menu if any error is logged
 - Added an option in settings to disable this marker (default: enabled to show it)
+- Added hint for Brizy users
 
 ### Changed